fix_4329 Modify K8S and docker mode for the performance test
