+++ ohara-it/src/test/scala/oharastream/ohara/it/performance/WithPerformanceRemoteConfigurator.scala
@@ -17,22 +17,18 @@
 package oharastream.ohara.it.performance
 
 import java.util.concurrent.TimeUnit
-
-import com.typesafe.scalalogging.Logger
 import oharastream.ohara.agent.DataCollie
 import oharastream.ohara.agent.container.ContainerClient
 import oharastream.ohara.agent.docker.DockerClient
 import oharastream.ohara.client.configurator.v0.NodeApi.{Node, State}
 import oharastream.ohara.common.util.{CommonUtils, Releasable, VersionUtils}
-import oharastream.ohara.it.{EnvTestingUtils, IntegrationTest, ServiceKeyHolder}
+import oharastream.ohara.it.{IntegrationTest, ServiceKeyHolder}
 import org.junit.{After, AssumptionViolatedException, Before}
 import org.scalatest.Matchers.contain
 import org.scalatest.Matchers._
 import scala.concurrent.ExecutionContext.Implicits.global
 
 abstract class WithPerformanceRemoteConfigurator extends IntegrationTest {
-  private[this] val log: Logger = Logger(classOf[WithPerformanceRemoteConfigurator])
-
   private[this] val CONFIURATOR_NODENAME_KEY = "ohara.it.performance.configurator.node"
 
   private[this] val configuratorNodeInfo: String = sys.env.getOrElse(
@@ -65,41 +61,14 @@ abstract class WithPerformanceRemoteConfigurator extends IntegrationTest {
   private[this] val configuratorContainerName: String =
     s"${configuratorContainerKey.group()}-${configuratorContainerKey.name()}"
 
-  private[this] val k8sURL: Option[String] =
-    sys.env.get("ohara.it.k8s").map(url => s"--k8s ${url}").orElse(Option.empty)
-
-  private[this] val serviceInfo: (Seq[Node], ContainerClient) = k8sURL
-    .map { argument =>
-      log.info(s"Running the K8S mode, K8S API argument is ${argument}")
-      val nodes: Seq[Node]                 = EnvTestingUtils.k8sNodes()
-      val containerClient: ContainerClient = EnvTestingUtils.k8sClient()
-      (nodes, containerClient)
-    }
-    .getOrElse {
-      log.info(s"Running the Docker mode")
-      val nodes: Seq[Node]                 = EnvTestingUtils.dockerNodes()
-      val containerClient: ContainerClient = DockerClient(DataCollie(nodes))
-      (nodes, containerClient)
-    }
-
   private[this] val imageName = s"oharastream/configurator:${VersionUtils.VERSION}"
 
-  protected val nodes: Seq[Node]                 = serviceInfo._1
-  protected val containerClient: ContainerClient = serviceInfo._2
+  protected def nodes: Seq[Node]
+  protected def containerClient: ContainerClient
 
   @Before
   def setupConfigurator(): Unit = {
     result(configuratorContainerClient.imageNames(configuratorHostname)) should contain(imageName)
-
-    val routes
-      : Map[String, String] = Map(configuratorNode.hostname -> CommonUtils.address(configuratorNode.hostname)) ++
-      nodes.map(node => node.hostname -> CommonUtils.address(node.hostname)).toMap ++
-      k8sURL
-        .map { url =>
-          val k8sMasterNodeName = url.split("http://").last.split(":").head
-          Map(k8sMasterNodeName -> CommonUtils.address(k8sMasterNodeName))
-        }
-        .getOrElse(Map.empty)
 
     result(
       configuratorContainerClient.containerCreator
@@ -107,10 +76,14 @@ abstract class WithPerformanceRemoteConfigurator extends IntegrationTest {
         .imageName(imageName)
         .portMappings(Map(configuratorPort -> configuratorPort))
         .command(
-          s"--hostname $configuratorHostname --port $configuratorPort ${k8sURL.getOrElse("")}"
+          s"--hostname $configuratorHostname --port $configuratorPort ${otherContainerCommand()}"
         )
         // add the routes manually since not all envs have deployed the DNS.
-        .routes(routes)
+        .routes(
+          Map(configuratorNode.hostname -> CommonUtils.address(configuratorNode.hostname)) ++
+            nodes.map(node => node.hostname -> CommonUtils.address(node.hostname)).toMap ++
+            otherRoutes()
+        )
         .name(configuratorContainerName)
         .create()
     )
@@ -118,6 +91,10 @@ abstract class WithPerformanceRemoteConfigurator extends IntegrationTest {
     // Wait configurator start completed
     TimeUnit.SECONDS.sleep(10)
   }
+
+  protected def otherRoutes(): Map[String, String] = Map.empty
+
+  protected def otherContainerCommand(): String = ""
 
   @After
   def releaseConfigurator(): Unit = {
